demand flex trans expansion and new grid sensitivity
