+++ PREPARE-TIMES-NZ/data_raw/coded_assumptions/commercial_demand/tech_fuel_capex.csv
@@ -7,8 +7,8 @@ Burner (Direct Heat),Coal,600,2023,Estimate based on relative boiler costs.,
 Burner (Direct Heat),LPG,300,2023,Estimate based on relative boiler costs.,
 Burner (Direct Heat),Natural Gas,300,2023,Similar to LPG above.,
 Burner (Direct Heat),Biogas,300,2023,Similar to LPG above.,
-Heat Pump (for Cooling),Electricity,1400,2022,,https://www.eeca.govt.nz/insights/eeca-insights/hot-water-heat-pumps-in-commercial-buildings/#:~:text=Capital%20cost
-Heat Pump (for Heating),Electricity,1400,2022,,https://www.eeca.govt.nz/insights/eeca-insights/hot-water-heat-pumps-in-commercial-buildings/#:~:text=Capital%20cost
+Heat Pump (for Cooling),Electricity,1200,2022,,https://www.eeca.govt.nz/insights/eeca-insights/hot-water-heat-pumps-in-commercial-buildings/#:~:text=Capital%20cost
+Heat Pump (for Heating),Electricity,1200,2022,,https://www.eeca.govt.nz/insights/eeca-insights/hot-water-heat-pumps-in-commercial-buildings/#:~:text=Capital%20cost
 Resistance Heater,Electricity,50,2025,,https://www.mitre10.co.nz/shop/heating-cooling/heating/electrical-heating/c/RS2059?srsltid=AfmBOopi8AEheKiqFv5ok30DAcHOxaq4gtjs0dCVd5aKjnqK7luqyedl&inStockNationwide=false&inStockSelectedStore=false&sort=RELEVANCY
 Direct Heat,Geothermal,150,2024,"High temperature geothermal resources in the Taup?, Rotorua and Kawerau districts deliver Geoheat across a range 
 of scales up to 10’s MWth and temperatures up to 220°C, with energy drawn from wells up to 3km deep. Estimated for 12 hour operation",https://www.nzgeothermal.org.nz/downloads/2024-2025-Geoheat-Action-Plan.pdf
