--- PREPARE-TIMES-NZ/data_raw/user_config/scenarios/LoadCurvesRipple.toml
@@ -1,29 +1,0 @@
-# Load Curves with Ripple
-# Scenario file to electricity load curves for each sector.
-# This alternative version shifts peak hot water demand to night timeslices 
-
-WorkBookName = "SuppXLS/Scen_LoadCurve_COM-FR_Rippled"
-
-[LoadCurvesCOMFR_IND]
-Description  = "Defines annual electricity demand timeslices or load curve for industrial sub-sectors"
-SheetName    = "IND"
-TagName      = "TFM_INS"
-DataLocation = "data_intermediate/stage_4_veda_format/scen_com_fr/com_fr_industry.csv"
-
-[LoadCurvesCOMFR_AGR]
-Description  = "Defines annual electricity demand timeslices or load curve for agriculture sub-sectors and processes"
-SheetName    = "AGR"
-TagName      = "TFM_INS"
-DataLocation = "data_intermediate/stage_4_veda_format/scen_com_fr/com_fr_agriculture.csv"
-
-[LoadCurvesCOMFR_COM]
-Description  = "Defines annual electricity demand timeslices or load curve for commercial sub-sectors"
-SheetName    = "COM"
-TagName      = "TFM_INS"
-DataLocation = "data_intermediate/stage_4_veda_format/scen_com_fr/com_fr_commercial.csv"
-
-[LoadCurvesCOMFR_RES]
-Description  = "Defines annual electricity demand timeslices or load curve for residential sub-sectors, including ripple hot water."
-SheetName    = "RES"
-TagName      = "TFM_INS"
-DataLocation = "data_intermediate/stage_4_veda_format/scen_com_fr/com_fr_residential_ripple.csv"
