+++ PREPARE-TIMES-NZ/data_raw/user_config/scenarios/sensitivities/low_transmission_costs.toml
@@ -1,0 +1,15 @@
+WorkBookName = "SuppXLS/Scen_TransmissionSensitivity"
+
+
+# A sensitivity scenario with 100 times cheaper transmission distribution 
+
+
+
+[DistributionSensitivityTest]
+Description = "Optional scenario file to reduce grid costs to 1% of original" 
+SheetName = "Distribution"  
+TagName = "FI_T"
+DataLocation = "data_intermediate/stage_4_veda_format/base_year_elc/distribution_parameters_sensitivity.csv"
+
+
+
